refactor login and signup
